renamed mount_wilson to hale
